requirements.txt
certifi
chardet
idna
requests
selenium
webdriver-manager
os
shutil
Pillow